+++ pybop/models/lithium_ion/cell_temperature.py
@@ -6,14 +6,15 @@ from pybamm import (
     Scalar,
     Variable,
 )
-from pybamm import lithium_ion as pybamm_lithium_ion
 from pybamm import t as pybamm_t
 from pybamm.models.full_battery_models.lithium_ion.electrode_soh import (
     get_min_max_stoichiometries,
 )
 
+from pybop.models.lithium_ion.base_model import BaseGroupedModel
 
-class CellTemperature(pybamm_lithium_ion.BaseModel):
+
+class CellTemperature(BaseGroupedModel):
     """
     A lumped thermal model for a battery cell.
 
@@ -157,16 +158,6 @@ class CellTemperature(pybamm_lithium_ion.BaseModel):
             out.print_name = r"U_\mathrm{p}(c^\mathrm{surf}_\mathrm{s,p})"
         return out
 
-    def build_model(self):
-        """
-        Build model variables and equations
-        Credit: PyBaMM
-        """
-        self._build_model()
-
-        self._built = True
-        pybamm.logger.info(f"Finish building {self.name}")
-
     @property
     def default_parameter_values(self) -> ParameterValues:
         param = ParameterValues("Chen2020")
@@ -284,55 +275,5 @@ class CellTemperature(pybamm_lithium_ion.BaseModel):
             "Heat transfer coefficient [W/K]": param["Heat transfer coefficient [W/K]"],
         }
         parameter_values = ParameterValues(values=parameter_dictionary)
-        parameter_values._set_initial_state = set_initial_state  # noqa: SLF001
+        parameter_values._set_initial_state = CellTemperature.set_initial_state  # noqa: SLF001
         return parameter_values
-
-
-def set_initial_state(
-    initial_value,
-    parameter_values,
-    direction=None,
-    param=None,
-    inplace=True,
-    options=None,
-    inputs=None,
-    tol=1e-6,
-):
-    """
-    Set the value of the initial state of charge.
-
-    Parameters
-    ----------
-    initial_value : float
-        Target initial value.
-        If float, interpreted as SOC, must be between 0 and 1.
-        If string e.g. "4 V", interpreted as voltage, must be between V_min and V_max.
-    parameter_values : :class:`pybamm.ParameterValues`
-        Parameters and their corresponding values.
-    param : :class:`pybamm.LithiumIonParameters`, optional
-        The symbolic parameter set to use for the simulation.
-        If not provided, the default parameter set will be used.
-    inplace: bool, optional
-        If True, replace the parameters values in place. Otherwise, return a new set of
-        parameter values. Default is True.
-    options : dict-like, optional
-        A dictionary of options to be passed to the model, see
-        :class:`pybamm.BatteryModelOptions`.
-    inputs : dict, optional
-        A dictionary of input parameters to pass to the model when solving.
-    tol : float, optional
-        The tolerance for the solver used to compute the initial stoichiometries.
-        A lower value results in higher precision but may increase computation time.
-        Default is 1e-6.
-    """
-    parameter_values = parameter_values if inplace else parameter_values.copy()
-
-    if isinstance(initial_value, int | float):
-        if not 0 <= initial_value <= 1:
-            raise ValueError("Initial SOC should be between 0 and 1")
-        parameter_values["Initial SoC"] = initial_value
-
-    else:
-        raise ValueError("Initial value must be a float between 0 and 1.")
-
-    return parameter_values